requirements.txt
tensorflow
opencv-python
Pillow
numpy
streamlit
matplotlib
pandas
